set notification values on input change

+++ src/main/java/com/adamk33n3r/runelite/watchdog/ui/nodegraph/NumberNodePanel.java
@@ -18,7 +18,7 @@ public class NumberNodePanel extends NodePanel {
 
         this.numValue = new ConnectionPointOut<>(this, node.getValue());
 
-        NumberInput numberInput = new NumberInput("Number Value", node.getValue().getValue().intValue());
+        NumberInput numberInput = new NumberInput("Number Value", node.getValue().getValue());
         this.items.add(new ConnectionLine<>(null, numberInput, this.numValue));
 
         this.pack();
